+++ scripts/run_cpp_example.py
@@ -73,7 +73,7 @@ class ExampleSpec:
 GUI_SCENE_EXAMPLE_DEFAULT_ARGS = {
     "imgui": (),
     "rigid_shapes": ("--scene", "mvp"),
-    "hello_world": ("--scene", "hello-world"),
+    "hello_world": (),
     "boxes": ("--scene", "boxes"),
     "box_stacking": ("--scene", "boxes"),
     "rigid_cubes": ("--scene", "boxes"),